% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON_file_processor.R
\name{compare_sa_ts}
\alias{compare_sa_ts}
\title{Compare the same models contained in two workspaces}
\usage{
compare_sa_ts(
  new_r_model = NA,
  new_model_workspace,
  old_model_workspace,
  materialized_ws_new = FALSE,
  materialized_ws_old = TRUE,
  java_processing_old_model = TRUE
)
}
\arguments{
\item{new_r_model}{-optional- an R model obtained via RJDemetra::get_model(workspace).
If != NA, this model is used as new model in the comparison.
Default = NA, new_model_workspace is used as new model}

\item{new_model_workspace}{the workspace (relative/absolute) path of the .xml
file or object used as new model in the comparison}

\item{old_model_workspace}{the workspace (relative/absolute) path of the .xml
file or object used as old model in the comparison}

\item{materialized_ws_new}{-optional- Default=FALSE, boolean field stating
whether the new workspace in the comparison is passed as
a matherialized or a virtual workspace (i.e. R object)}

\item{materialized_ws_old}{-optional- Default=TRUE, boolean field stating
whether the new workspace in the comparison is passed as
a matherialized or a virtual workspace (i.e. R object)}

\item{java_processing_old_model}{-optional- Default=TRUE, use Java models for
the internal processing of the old workspace (faster than R)}
}
\value{
a "comparisons.pdf" file containig the plots of the seasonally adjusted
time series of both the new and old workspaces. Series with the same
series_name are reported in the same plot
}
\description{
This function compares the same models contained in two workspaces (old and new)
plotting the respective sasonally adjusted series into a pdf file
}
\examples{
require(RJDemetra)
original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
input_workspace_materialized <- "workspace_test/workspace.xml"
virtual_workspace <- load_workspace(input_workspace_materialized)
compute(virtual_workspace)
compare_sa_ts(new_model_workspace = virtual_workspace,
              old_model_workspace = input_workspace_materialized ,
               materialized_ws_new=FALSE, materialized_ws_old=TRUE,
               java_processing_old_model=FALSE)
setwd(original_directory)
}
