% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_manager.R
\name{compare_workspaces}
\alias{compare_workspaces}
\title{Compare two RJDemetra workspaces}
\usage{
compare_workspaces(
  workspace1,
  workspace2,
  output_file = "differences_report.txt"
)
}
\arguments{
\item{workspace1}{Either an RJDemetra workspace object or a file path to a workspace.}

\item{output_file}{Character string specifying the path of the output report file.
Default is "differences_report.txt".}

\item{patworkspace2}{Either an RJDemetra workspace object or a file path to a workspace.}
}
\value{
The function does not return a value but writes a report to the specified file.
}
\description{
This function compares two RJDemetra workspaces, either provided as objects
or file paths, and generates a report of the differences in the final
seasonally adjusted series and likelihood BIC values.
}
\details{
The function loads the workspaces if they are provided as file paths,
computes the models, and extracts the seasonally adjusted series and likelihood BIC.
It then compares these values and records any differences in the output file.

Differences are checked for the last 36 observations of the seasonally adjusted series
and the likelihood BIC values.
}
