% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convert_numeric_matrix_to_mts}
\alias{convert_numeric_matrix_to_mts}
\title{Convert a numeric matrix to an mts object}
\usage{
convert_numeric_matrix_to_mts(data_matrix, freq = NULL)
}
\arguments{
\item{data_matrix}{A numeric matrix where row names are dates in "YYYY-MM-DD" format,
and column names are the names of the time series.}

\item{freq}{-optional- The frequency of the time series. If not provided, the function will
try to infer the frequency based on the difference between the first two dates.}
}
\value{
An mts object with the given time series data.
}
\description{
This function takes a numeric matrix with dates as row names and series names
as column names, and converts it to an mts (multivariate time series) object.
It is goot for converting data read by data_readers and ext_reg_data_readers
into mts that could be useful as input as usrdef.var for RJDemetra functions
tramoseats() and X13()
}
\examples{
input_data_file_name <- system.file("extdata","CSV-FAS/grezzi_trim_FAS.csv", package = "RJDProcessor")
input_data_reader    <- Data_reader_csv(input_source = input_data_file_name)
data_matrix <- input_data_reader@read_data()

# Convert the matrix to an mts object
mts_object <- convert_numeric_matrix_to_mts(data_matrix)
print(mts_object)
}
