% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_csv.R
\name{Data_reader_csv}
\alias{Data_reader_csv}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_csv(input_source = NA, ...)
}
\arguments{
\item{input_source}{A string with file name (also with path).}
}
\value{
The Data_reader_csv object
}
\description{
This function creates a Data_reader object capable of reading data from CSV files and returning it using the \code{read_data()} function.
}
\examples{
input_data_file_name <- "CSV-FAS/grezzi_trim_FAS.csv"
input_data_reader <- Data_reader_csv(input_source = input_data_file_name)
#input_data_readerATread_data() # uncomment and replace AT with its symbol
}
