% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_csv_istat_format.R
\name{Data_reader_csv_istat_format}
\alias{Data_reader_csv_istat_format}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_csv_istat_format(input_source = NA, ...)
}
\arguments{
\item{input_source}{A string with file name (also with path).}
}
\value{
The Data_reader_csv_istat_format object
}
\description{
This function creates a Data_reader object capable of reading data from CSV files in ISTAT format and returning it using the \code{read_data()} function.
The ISTAT format is a csv file with dates in format YYYYqMM as rownames and time_series names as colnames
}
\examples{
input_data_file_name <- system.file("extdata","SITIC-TUR/grezziTUR.csv", package = "RJDProcessor")
input_data_reader <- Data_reader_csv_istat_format(input_source = input_data_file_name)
#input_data_reader@read_data()
}
