% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_ext_reg_csv.R
\name{Data_reader_ext_reg_csv}
\alias{Data_reader_ext_reg_csv}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_ext_reg_csv(input_source, ...)
}
\arguments{
\item{input_source}{A string with the input: e.g. a file name (also with path) if the input is a file.}
}
\value{
The Data_reader_ext_reg_csv object
}
\description{
This function creates a Data_reader_ext_reg object capable of reading data from CSV external regressors files and returning it using the \code{read_ext_reg_data()} function.
}
\examples{
require(RJDemetra)
input_workspace_xml       <- system.file("extdata", "WorkspaceTUR-container/workspace-TUR.xml",
                                         package = "RJDProcessor")
input_data_file_name      <- system.file("extdata", "CSV-TUR/grezzi_trim_TUR.csv", package = "RJDProcessor")
regr_directory            <- system.file("extdata", "CSV-TUR/regr", package = "RJDProcessor")
ws                        <- load_workspace(file = input_workspace_xml)
compute(ws)
data_reader_ext_reg       <- Data_reader_ext_reg_csv(regr_directory)
all_model_ext_vars_info <- data_reader_ext_reg@read_ext_reg_info(ws)
vars_matrix             <- data_reader_ext_reg@read_ext_reg_data(all_model_ext_vars_info, "VATASC",
                                                                 frequency=12)

}
