% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_list.R
\name{Data_reader_list}
\alias{Data_reader_list}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_list(input_source = NA, ...)
}
\arguments{
\item{input_source}{A string with file name (also with path).}
}
\value{
The Data_reader_csv object
}
\description{
This function creates a Data_reader object capable of reading data from a list and returning it using the \code{read_data()} function.
}
\examples{
FATEXP_10_list <- list("series_name"="FATEXP_10", "dates"=c("2005-01-01","2005-02-01","2005-03-01"),
                       "values"=c(12, 15, 11.1))
C_DEFL_list    <- list("series_name"="C_DEFL",    "dates"=c("2001-01-01","2001-02-01","2001-03-01"),
                       "values"=c(99, 100, 99.1))
# ...
input_data_list   <- list(FATEXP_10_list, C_DEFL_list) #add more time series if you want (here are 2)
input_data_reader <- Data_reader_list(input_source = input_data_list)
input_data_reader@read_data()
}
