% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_xlsx.R
\name{Data_reader_xlsx}
\alias{Data_reader_xlsx}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_xlsx(input_source = NA, ...)
}
\arguments{
\item{input_source}{A string with file name (also with path).}
}
\value{
The Data_reader_xlsx object
}
\description{
This function creates a Data_reader object capable of reading data from XLSX files and returning it using the \code{read_data()} function.
}
\examples{
input_data_file_name <- system.file("extdata","XLSX-TUR/grezzi_trim_TUR.xlsx", package = "RJDProcessor")
input_data_reader    <- Data_reader_xlsx(input_source = input_data_file_name)
input_data_reader@read_data()
}
