% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_xml.R
\name{Data_reader_xml}
\alias{Data_reader_xml}
\title{Constructor (R-like) of the Data_reader object}
\usage{
Data_reader_xml(input_source = NA, ...)
}
\arguments{
\item{input_source}{A string with file name (also with path).}
}
\value{
The Data_reader_xml object
}
\description{
This function creates a Data_reader object capable of reading data from XLSX files and returning it using the \code{read_data()} function.
}
\examples{
input_data_file_name <- system.file("extdata","Prod.xml", package = "RJDProcessor")
# NOTE: absolute paths are better for this Data_reader
input_data_reader    <- Data_reader_xml(input_source = input_data_file_name)
#input_data_reader@read_data() # for reading the data
}
