% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON.R
\name{from_reduced_to_full_JD_JSON_file}
\alias{from_reduced_to_full_JD_JSON_file}
\title{Print a JSON file with all the JD_JSON fields explicit}
\usage{
from_reduced_to_full_JD_JSON_file(
  JD_JSON_file,
  output_file_name = NA,
  indent = TRUE
)
}
\arguments{
\item{JD_JSON_file}{The name of the file in which the JD_JSON will be saved}

\item{output_file_name}{-optional- The name of the file (optionally with path) in
which the JD_JSON will be saved. If NA, the name of the
file will be paste0(JD_DSON_file,"_full")}

\item{indent}{-optional- Default TRUE. Print each field of the JSON in a
different row}
}
\value{
Void. A JSON file is saved on the filesystem with all the JD_JSON fields
}
\description{
This function prints a JSON string that contains all the fields of the JD_JSON object explicitly defined.
}
\examples{
original_directory <- getwd()
extdata_directory <- system.file("extdata", package = "RJDProcessor")
JSON_file_name <- system.file("extdata", "specifications_example2.txt", package = "RJDProcessor")
setwd(extdata_directory)
from_reduced_to_full_JD_JSON_file(JSON_file_name)
setwd(original_directory)
}
