% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON_file_processor.R
\name{get_r_model_from_j_model}
\alias{get_r_model_from_j_model}
\title{Get an R list with model information from java model}
\usage{
get_r_model_from_j_model(j_model)
}
\arguments{
\item{j_model}{a Java model obtained from RJDemetra}
}
\value{
a list containing all the models specifications
}
\description{
This function gets an R list with model information from java model
}
\examples{
require(RJDemetra)
original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
input_workspace <- "workspace_test/workspace.xml"
virtual_workspace <- load_workspace(input_workspace)
compute(virtual_workspace)
m <- get_jmodel(virtual_workspace)
r_model_list <- get_r_model_from_j_model(m)
setwd(original_directory)
}
