% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_manager.R
\name{get_single_ts_workspaces}
\alias{get_single_ts_workspaces}
\title{Get a single workspace for each series in a workspace}
\usage{
get_single_ts_workspaces(
  full_workspace,
  single_workspaces_path,
  compressed_ws = TRUE,
  clean_single_ws_directory = TRUE,
  from_TS_PLUS_plugin = TRUE
)
}
\arguments{
\item{full_workspace}{A workspace with multiple time series SA models}

\item{single_workspaces_path}{The path in which the single workspcaes will be
stored}

\item{compressed_ws}{-optional- Default=TRUE, compress the single workspaces
that will be created}

\item{clean_single_ws_directory}{-optional- Default=TRUE, delete the original
workspace after creating the single series ones}

\item{from_TS_PLUS_plugin}{-optional- Default=TRUE, the original workspace is
created by the SA_ext_plugin. In that case some internal
adjustments are needed and they could make the computation
slower. There are no problems if this field is TRUE but
the ws is created without SA_ext_plugin}
}
\value{
A set of materialized workspaces in single_workspaces_path, one workspace
for each time series SA model contained in full_workspace
}
\description{
This function gets a single time series workspace for each time series SA model
present in a given workspace
}
\examples{
require(RJDemetra)
original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
single_workspaces_path <- "splitted_workspaces"
full_workspace         <- "WorkspaceTUR-container\\\\workspace-TUR.xml"
compressed_ws          <-  FALSE
get_single_ts_workspaces(full_workspace, single_workspaces_path,
                         compressed_ws = compressed_ws )
setwd(original_directory)
}
