% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON.R
\name{JD_JSON_to_materialized_workspace}
\alias{JD_JSON_to_materialized_workspace}
\title{Turn a JD_JSON in a materialized workspace}
\usage{
JD_JSON_to_materialized_workspace(
  workspace_dir = NA,
  JSON_file,
  input_data_reader,
  ext_reg_data_reader = NA,
  series_to_proc_names = NA
)
}
\arguments{
\item{workspace_dir}{-optional- the directory of the input workspace.
Default = NA stores the workspace in a directory called "output_workspace_container"}

\item{JSON_file}{Name of the JSON file that will be produced (also with path).}

\item{input_data_reader}{A Data_Reader object}

\item{ext_reg_data_reader}{-optional- A Data_reader_ext_reg object, to produce the metadata of the external regressors
(i.e. create the names of the csv rather than a xlsx files o other containers for external regressors)
Default = NA, do not consider external regressors (discouraged)}

\item{series_to_proc_names}{-optional- an array containing the name of the series to be included in the workspace
among the ones present in the JD_JSON file e.g. c("VATASA","VATAIA")}
}
\value{
void in R environment, a workspace materialized in the filesystem
}
\description{
This function obtain a JD_JSON file from a workspace stored in the filesystem (in a directory).
See test foder for examples
}
\examples{
require(RJDemetra)
input_JD_JSON        <- system.file("extdata", "specifications_example1.txt", package = "RJDProcessor")
input_data_file_name <- system.file("extdata", "CSV-TUR/grezzi_trim_TUR.csv", package = "RJDProcessor")
regr_directory       <- system.file("extdata", "CSV-TUR/regr", package = "RJDProcessor")
input_data_reader         <- Data_reader_csv(input_source = input_data_file_name)
ext_reg_input_data_reader <- Data_reader_ext_reg_csv(regr_directory)
JD_JSON_to_materialized_workspace(workspace_dir="ws_out" ,input_JD_JSON, input_data_reader, ext_reg_input_data_reader, series_to_proc_names=NA)

}
