% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON.R
\name{JD_JSON_to_virtual_workspace}
\alias{JD_JSON_to_virtual_workspace}
\title{Turn a JD_JSON in a virtual workspace}
\usage{
JD_JSON_to_virtual_workspace(
  JSON_file,
  input_data_reader,
  ext_reg_data_reader = NA,
  series_to_proc_names = NA
)
}
\arguments{
\item{JSON_file}{Name of the JSON file to turn in a workspace (also with path).}

\item{input_data_reader}{A Data_Reader object}

\item{ext_reg_data_reader}{-optional- A Data_reader_ext_reg object, to read the external regressors
in the desired format (csv, xlsx, tramoseats+, ...)
Default = NA, do not consider external regressors (discouraged)}

\item{series_to_proc_names}{-optional- an array containing the name of the series to be included in the workspace
among the ones present in the JD_JSON file e.g. c("VATASA","VATAIA")}
}
\value{
A virtual workspace
}
\description{
This function obtain a virtual workspace from a JD_JSON file.
See test foder for examples
}
\examples{
require(RJDemetra)
input_JD_JSON        <- system.file("extdata", "specifications_example1.txt", package = "RJDProcessor")
input_data_file_name <- system.file("extdata", "CSV-TUR/grezzi_trim_TUR.csv", package = "RJDProcessor")
regr_directory       <- system.file("extdata", "CSV-TUR/regr", package = "RJDProcessor")
input_data_reader         <- Data_reader_csv(input_source = input_data_file_name)
ext_reg_input_data_reader <- Data_reader_ext_reg_csv(regr_directory)
ws <- JD_JSON_to_virtual_workspace(input_JD_JSON, input_data_reader, ext_reg_input_data_reader, series_to_proc_names=NA)

}
