% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_manager.R
\name{merge_workspaces}
\alias{merge_workspaces}
\title{Merge many workspaces in one}
\usage{
merge_workspaces(
  source_workspaces_path,
  merged_ws_name = "merged_ws",
  merged_ws_dir = NA,
  compressed = TRUE,
  delete_originals = TRUE,
  silent = TRUE
)
}
\arguments{
\item{source_workspaces_path}{The path in which to find all the workspcaes that
will be merged}

\item{merged_ws_name}{-optional- Default="merged_ws". The name of the workspace
that will be created}

\item{merged_ws_dir}{-optional- Default=NA, the path in which the merged workspace
will be stored. If NA it will be stored in the current
directory}

\item{compressed}{-optional- Default=TRUE, workspaces to be merged are
compressed}

\item{delete_originals}{-optional- Default=TRUE, delete original workspaces
after merging them}

\item{silent}{-optional- Default=TRUE, do not print status bar and messages
during the operations}
}
\value{
A virtual (R) and a marerialized workspace containing all the
multiprocessings and time series SA models of the original workspaces
}
\description{
This function gets merges many workspaces contained in a given folder into one
workspace
}
\examples{
require(RJDemetra)
original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
single_workspaces_path <- "splitted_workspaces_to_merge"
full_workspace_path    <- "merged_ws_test"
compressed_ws          <-  FALSE
ws_merged <- merge_workspaces(source_workspaces_path=single_workspaces_path,
             merged_ws_dir = full_workspace_path, compressed = TRUE,
             delete_originals = FALSE, silent=TRUE)

setwd(original_directory)
}
