% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JD_JSON_file_processor.R
\name{produce_fully_R_workspace}
\alias{produce_fully_R_workspace}
\title{Process a JD_JSON file}
\usage{
produce_fully_R_workspace(
  ext_reg_input_data_reader = NA,
  original_ws_xml,
  new_r_ws_folder
)
}
\arguments{
\item{ext_reg_input_data_reader}{A specific Data_reader_ext_reg object
(CSV, XLSX, ...) to read the external
regressors. Default=NA.}

\item{original_ws_xml}{xml file (also with path) of the workspace to be
transformed into a fully R workspace}

\item{new_r_ws_folder}{Name of the directory that will contain
the output, fully R, workspace.}
}
\value{
a virtual workspace, already processed
}
\description{
This function processes a JSON file with JD_JSON fields and returns a virtual
workspace
}
\examples{
require(RJDemetra)
original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
original_ws_xml <- "WorkspaceTUR-container/workspace-TUR.xml"
new_r_ws_folder <- "output_fully_R_ws"
regr_directory  <- "SITIC-TUR/regr"
ext_reg_input_data_reader <- Data_reader_ext_reg_tsplus(regr_directory)
produce_fully_R_workspace(ext_reg_input_data_reader, original_ws_xml, new_r_ws_folder)


extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
original_ws_xml <- "WorkspaceFAS-standard-container//FAS.xml"
new_r_ws_folder <- "output_fully_R_ws"
regr_directory  <- "CSV-FAS//regr"
ext_reg_input_data_reader <- Data_reader_ext_reg_csv(regr_directory)
produce_fully_R_workspace(ext_reg_input_data_reader, original_ws_xml, new_r_ws_folder)
setwd(original_directory)

}
