% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_csv.R
\name{read_data,Data_reader_csv-method}
\alias{read_data,Data_reader_csv-method}
\title{Get the data from a Data_reader_csv}
\usage{
\S4method{read_data}{Data_reader_csv}(object, ...)
}
\value{
data in form of numeric matrix, with rownames = dates (in string format, YYYY-MM-DD) end colnames = time series names (string)
}
\description{
This function returns the data from the input_source of the object.
}
\examples{
input_data_file_name <- "CSV-FAS/grezzi_trim_FAS.csv"
input_data_reader <- Data_reader_csv(input_source = input_data_file_name)
#input_data_readerATread_data() # uncomment and replace AT with its symbol
}
