% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_list.R
\name{read_data,Data_reader_list-method}
\alias{read_data,Data_reader_list-method}
\title{Get the data from a Data_reader_list}
\usage{
\S4method{read_data}{Data_reader_list}(object, ...)
}
\value{
data in form of numeric matrix, with rownames = dates (in string format, YYYY-MM-DD) and colnames = time series names (string)
}
\description{
This function returns the data from the input_source of the object.
}
\examples{

FATEXP_10_list <- list("series_name"="FATEXP_10", "dates"=c("2005-01-01","2005-02-01","2005-03-01"),
                       "values"=c(12, 15, 11.1))
C_DEFL_list    <- list("series_name"="C_DEFL",    "dates"=c("2001-01-01","2001-02-01","2001-03-01"),
                       "values"=c(99, 100, 99.1))
#...
input_data_list   <- list(FATEXP_10_list, C_DEFL_list) #add more time series if you want (here are 2)
input_data_reader <- Data_reader_list(input_source = input_data_list)
input_data_reader@read_data()
}
