% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_workspace.R
\name{read_data,Data_reader_workspace-method}
\alias{read_data,Data_reader_workspace-method}
\title{Get the data from a Data_reader_workspace}
\usage{
\S4method{read_data}{Data_reader_workspace}(object, ...)
}
\value{
data in form of numeric matrix, with rownames = dates (in string format, YYYY-MM-DD) and colnames = time series names (string)
}
\description{
This function returns the data from the input_source of the object.
}
\examples{
input_data_file_name <- system.file("extdata","WorkspaceTUR-container/workspace-TUR.xml", package = "RJDProcessor")
input_data_reader    <- Data_reader_workspace(input_source = input_data_file_name)
mts <- input_data_reader@read_data()
}
