% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_xlsx.R
\name{read_data,Data_reader_xlsx-method}
\alias{read_data,Data_reader_xlsx-method}
\title{Get the data from a Data_reader_csv}
\usage{
\S4method{read_data}{Data_reader_xlsx}(object, ...)
}
\value{
data in form of numeric matrix, with rownames = dates (in string format, YYYY-MM-DD) and colnames = time series names (string)
}
\description{
This function returns the data from the input_source of the object.
}
\examples{
input_data_file_name <- system.file("extdata","XLSX-TUR/grezzi_trim_TUR.xlsx", package = "RJDProcessor")
input_data_reader    <- Data_reader_xlsx(input_source = input_data_file_name)
input_data_reader@read_data()
}
