% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_xml.R
\name{read_data,Data_reader_xml-method}
\alias{read_data,Data_reader_xml-method}
\title{Get the data from a Data_reader_xml}
\usage{
\S4method{read_data}{Data_reader_xml}(object, ...)
}
\value{
data in form of numeric matrix, with rownames = dates (in string format, YYYY-MM-DD) and colnames = time series names (string)
}
\description{
This function returns the data from the input_source of the object.
}
\examples{
input_data_file_name <- system.file("extdata","Prod.xml", package = "RJDProcessor")
# NOTE: absolute paths are better for this Data_reader
input_data_reader    <- Data_reader_xml(input_source = input_data_file_name)
#input_data_reader@read_data() # for reading the data
}
