% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_ext_reg_csv.R
\name{read_ext_reg_data,Data_reader_ext_reg_csv-method}
\alias{read_ext_reg_data,Data_reader_ext_reg_csv-method}
\title{Read external regressors data}
\usage{
\S4method{read_ext_reg_data}{Data_reader_ext_reg_csv}(
  object,
  var_info = NULL,
  time_series_info = NULL,
  frequency = NA_integer_,
  ...
)
}
\arguments{
\item{var_info}{A string with file name (also with path).}

\item{time_series_info}{A string with time series name in workspace name (also with path).}

\item{frequency}{i.e. 12 = monthly data, 4 = quarterly data}
}
\value{
a numeric matrix with variable names as colnames and YYYY-MM-DD dates as rownames
}
\description{
This function reads data from external regressors and returns it as a numeric matrix with variable names as colnames
and YYYY-MM-DD dates as rownames
}
\examples{
require(RJDemetra)
input_workspace_xml       <- system.file("extdata", "WorkspaceTUR-container/workspace-TUR.xml",
                                         package = "RJDProcessor")
input_data_file_name      <- system.file("extdata", "CSV-TUR/grezzi_trim_TUR.csv", package = "RJDProcessor")
regr_directory            <- system.file("extdata", "CSV-TUR/regr", package = "RJDProcessor")
ws                        <- load_workspace(file = input_workspace_xml)
compute(ws)
data_reader_ext_reg       <- Data_reader_ext_reg_csv(regr_directory)
all_model_ext_vars_info <- data_reader_ext_reg@read_ext_reg_info(ws)
vars_matrix             <- data_reader_ext_reg@read_ext_reg_data(all_model_ext_vars_info, "VATASC",
                                                                 frequency=12)

}
