% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_reader_ext_reg_xlsx.R
\name{read_ext_reg_info,Data_reader_ext_reg_xlsx-method}
\alias{read_ext_reg_info,Data_reader_ext_reg_xlsx-method}
\title{Read information about external regressors from a workspace}
\usage{
\S4method{read_ext_reg_info}{Data_reader_ext_reg_xlsx}(object, var_info_container, adjust_path = TRUE, ...)
}
\arguments{
\item{var_info_container}{workspace xml file path}
}
\value{
list() of information about external regressors
}
\description{
This function returns a list of information about external regressors used in the models contained in a workspaces
}
\examples{

require(RJDemetra)
input_workspace_xlsx      <- system.file("extdata", "WorkspaceTUR-container/workspace-TUR.xml", package = "RJDProcessor")
input_data_file_name      <- system.file("extdata", "XLSX-TUR/grezzi_trim_TUR.xlsx", package = "RJDProcessor")
regr_directory            <- system.file("extdata", "XLSX-TUR/regr", package = "RJDProcessor")
ws                        <- load_workspace(file = input_workspace_xlsx)
compute(ws)
data_reader_ext_reg       <- Data_reader_ext_reg_xlsx(regr_directory)
all_model_ext_vars_info <- data_reader_ext_reg@read_ext_reg_info(ws)

}
