% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{seasonality_changes_report}
\alias{seasonality_changes_report}
\title{Report dei cambiamenti nella stagionalità tra due workspace}
\usage{
seasonality_changes_report(
  workspace_old,
  workspace_new,
  output_file = "seasonality_changes.txt"
)
}
\arguments{
\item{workspace_old}{Oggetto workspace di RJDemetra precedente.}

\item{workspace_new}{Oggetto workspace di RJDemetra attuale.}

\item{output_file}{Percorso del file .txt dove verrà salvato il report.}
}
\value{
La funzione non restituisce oggetti in R, ma scrive un file di testo.
}
\description{
Confronta i modelli ARIMA di due workspace (vecchio e nuovo) e identifica le serie
che hanno cambiato il loro stato di stagionalità (passando da modelli stagionali a non stagionali e viceversa).
}
