% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_diagnostic_report1}
\alias{create_diagnostic_report1}
\title{Create Diagnostic Report for Time Series Models}
\usage{
create_diagnostic_report1(workspace, output_file = "report.txt")
}
\arguments{
\item{workspace}{A workspace object containing the time series models.}

\item{output_file}{A string specifying the output file path where the report will be saved (default is "report.txt").}
}
\description{
This function generates a diagnostic report for time series models,
including BIC, p-values from Ljung-Box tests, and normality test results.
It organizes the output into categories: all series, series with Ljung-Box problems,
series with normality issues, and a summary of problematic series.
}
\examples{
require(RJDemetra)

original_directory <- getwd()
extdata_directory  <- system.file("extdata", package = "RJDProcessor")
setwd(extdata_directory)
ws_path <- "WorkspaceTUR-container/workspace-TUR.xml"

workspace <- load_workspace(file=ws_path)
create_diagnostic_report1(workspace, output_file = "report.out")
setwd(original_directory)
}
